support non-integral refresh intervals, possibly as low as 0.1s. this makes it more useful for real time display of fast changing stats ok art deraadt

+++ usr.bin/systat/extern.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: extern.h,v 1.16 2007/02/23 22:02:58 deanna Exp $	*/
+/*	$OpenBSD: extern.h,v 1.17 2007/02/24 09:45:27 otto Exp $	*/
 /*	$NetBSD: extern.h,v 1.3 1996/05/10 23:16:34 thorpej Exp $	*/
 
 /*-
@@ -48,7 +48,7 @@ extern int	*dk_select;
 extern int	CMDLINE;
 extern int	dk_ndrive;
 extern int	hz, stathz;
-extern u_int	naptime;
+extern double	naptime;
 extern int	nhosts;
 extern int	nports;
 extern int	protos;
